fixup! feat: implement logic for different list views with routing logic
