+++ components/ModEditSubmissionTopbar.vue
@@ -50,9 +50,9 @@
 import { ref, type Ref } from 'vue'
 import SVGCopyContent from '~/assets/icons/content-copy.svg'
 import SVGSuccessCheckMark from '~/assets/icons/checkmark-square.svg'
-import { useModerationSubmissionsListStore } from '~/stores/moderationSubmissionsListStore'
+import { useModerationSubmissionsStore } from '~/stores/moderationSubmissionsStore'
 
-const modSubmissionStore = useModerationSubmissionsListStore()
+const modSubmissionStore = useModerationSubmissionsStore()
 const selectedSubmissionId: Ref<string> = ref(modSubmissionStore.selectedSubmissionId)
 
 const showCopySuccessIcon: Ref<boolean> = ref(false)
